+++ examples/react-template/screens/Card.tsx
@@ -84,11 +84,8 @@ export const CardScreen = (): JSX.Element => {
       <Title level={TitleLevels.TWO}>CardImage & markup </Title>
       <Spacer size={SpacerSize.FOUR} />
       <Card>
-        <CardImage src='https://i.etsystatic.com/10951167/r/il/df66c4/1860902191/il_570xN.1860902191_kuoj.jpg' />
         <CardContent>
-          <Title overline level={TitleLevels.FIVE}>
-            Markup {'<p>'} - level FIVE
-          </Title>
+          <Title level={TitleLevels.FIVE}>Markup {'<p>'} - level FIVE</Title>
           <Text>
             Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed ligula ex, aliquam at neque eu, vulputate vera.
           </Text>
@@ -96,16 +93,14 @@ export const CardScreen = (): JSX.Element => {
             Skeleton toogle click
           </Button>
         </CardContent>
+        <CardImage src='https://i.etsystatic.com/10951167/r/il/df66c4/1860902191/il_570xN.1860902191_kuoj.jpg' />
       </Card>
 
       <Title level={TitleLevels.TWO}>Reversed </Title>
       <Spacer size={SpacerSize.FOUR} />
-      <Card>
-        <CardImage src='https://i.etsystatic.com/10951167/r/il/df66c4/1860902191/il_570xN.1860902191_kuoj.jpg' />
+      <Card reversed>
         <CardContent>
-          <Title overline level={TitleLevels.FIVE}>
-            Markup {'<p>'} - level FIVE
-          </Title>
+          <Title level={TitleLevels.FIVE}>Markup {'<p>'} - level FIVE</Title>
           <Text>
             Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed ligula ex, aliquam at neque eu, vulputate vera.
           </Text>
@@ -113,12 +108,12 @@ export const CardScreen = (): JSX.Element => {
             Skeleton toogle click
           </Button>
         </CardContent>
+        <CardImage src='https://i.etsystatic.com/10951167/r/il/df66c4/1860902191/il_570xN.1860902191_kuoj.jpg' />
       </Card>
 
       <Title level={TitleLevels.TWO}>Flat </Title>
       <Spacer size={SpacerSize.THREE} />
       <Card flat>
-        <CardImage src='https://i.etsystatic.com/10951167/r/il/df66c4/1860902191/il_570xN.1860902191_kuoj.jpg' />
         <CardContent>
           <Title level={TitleLevels.THREE}>Markup {'<p>'} - level THREE</Title>
           <Text>
@@ -131,16 +126,13 @@ export const CardScreen = (): JSX.Element => {
             Enabled
           </Button>
         </CardContent>
+        <CardImage src='https://i.etsystatic.com/10951167/r/il/df66c4/1860902191/il_570xN.1860902191_kuoj.jpg' />
       </Card>
 
       <Title level={TitleLevels.TWO}>Horizontal </Title>
       <Spacer size={SpacerSize.THREE} />
       <Spacer size={SpacerSize.THREE} />
       <Card horizontal>
-        <CardImage
-          size={'3'}
-          src='https://i.etsystatic.com/10951167/r/il/df66c4/1860902191/il_570xN.1860902191_kuoj.jpg'
-        />
         <CardContent>
           <Title level={TitleLevels.FIVE} markup='h5'>
             Markup H5 - level FOUR
@@ -155,6 +147,33 @@ export const CardScreen = (): JSX.Element => {
             Enabled
           </Button>
         </CardContent>
+        <CardImage
+          size={'3'}
+          src='https://i.etsystatic.com/10951167/r/il/df66c4/1860902191/il_570xN.1860902191_kuoj.jpg'
+        />
+      </Card>
+
+      <Title level={TitleLevels.TWO}>Horizontal reversed </Title>
+      <Spacer size={SpacerSize.SIX} />
+      <Card horizontal reversed>
+        <CardContent>
+          <Title level={TitleLevels.FIVE} markup='h5'>
+            Markup H5 - level FOUR
+          </Title>
+          <Text>
+            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed ligula ex, aliquam at neque eu, vulputate vera.
+          </Text>
+          <Button
+            variant={ButtonVariant.PRIMARY}
+            onClick={() => window.open('https://fr.wikipedia.org/wiki/Shiba_(chien)', '_blank')}
+          >
+            Enabled
+          </Button>
+        </CardContent>
+        <CardImage
+          size={'1'}
+          src='https://i.etsystatic.com/10951167/r/il/df66c4/1860902191/il_570xN.1860902191_kuoj.jpg'
+        />
       </Card>
 
       <Title level={TitleLevels.TWO}>Floating (inside component)</Title>
@@ -169,7 +188,6 @@ export const CardScreen = (): JSX.Element => {
               </AccordionHeader>
               <AccordionBody>
                 <Card floating>
-                  <CardImage src='https://i.etsystatic.com/10951167/r/il/df66c4/1860902191/il_570xN.1860902191_kuoj.jpg' />
                   <CardContent>
                     <Text>Exemple de Card sans titre</Text>
                     <Button
@@ -179,6 +197,7 @@ export const CardScreen = (): JSX.Element => {
                       Enabled
                     </Button>
                   </CardContent>
+                  <CardImage src='https://i.etsystatic.com/10951167/r/il/df66c4/1860902191/il_570xN.1860902191_kuoj.jpg' />
                 </Card>
               </AccordionBody>
             </AccordionItem>