+++ packages/react/components/autocomplete/AutoComplete.tsx
@@ -26,7 +26,7 @@ const AutoCompleteRef = <T extends string | Item<unknown> = string>(
     onChange,
     name,
     matching = defaultMatching,
-    displayMenu = true,
+    displayMenu,
     onItemSelected,
     iconNameLeft,
     iconNameRight,
@@ -39,6 +39,7 @@ const AutoCompleteRef = <T extends string | Item<unknown> = string>(
     onFocus,
     id,
     loading,
+    ...others
   }: AutoCompletePropsWeb<T>,
   ref: React.Ref<HTMLInputElement>,
 ): JSX.Element => {
@@ -91,6 +92,7 @@ const AutoCompleteRef = <T extends string | Item<unknown> = string>(
         onChange={handleChange}
         onIconClick={onIconClick}
         loading={loading}
+        {...others}
       />
 
       {isAutocompleteMenuVisible && (
@@ -111,7 +113,6 @@ const AutoCompleteRef = <T extends string | Item<unknown> = string>(
           )}
         </div>
       )}
-      <div />
     </div>
   )
 }