+++ packages/react/components/input/Input.tsx
@@ -164,7 +164,7 @@ const Input = React.forwardRef<InputRef, InputProp>(
     return (
       <div className={wrapperClasses} data-has-gauge={securityGauge ? true : undefined}>
         {label && (
-          <label className='input-label' htmlFor={id}>
+          <label className={hashClass('input-label')} htmlFor={id}>
             {label}{' '}
             {required && (
               <Text markup={TextMarkup.SPAN} typo={TypographyColor.TEXT_ERROR}>