+++ packages/react/components/pagination/Pagination.native.tsx
@@ -9,7 +9,7 @@ import { Text } from '@/components/text'
 
 /**
  * Pagination Component
- * @param count {number} Number elements
+ * @param length {number} Number of pages
  * @param defaultPage {number} Current default active page (default is 1)
  * @param pageSize {number} Element per page (default is 10)
  * @param onClick {Function} Return pagination object