+++ packages/react/components/pagination/PaginationProps.ts
@@ -10,7 +10,6 @@ import { View } from 'react-native'
 export interface PaginationProps extends Accessibility, Dev, CommonProps {
   length: number
   defaultPage?: number
-  pageSize?: number
   onClick?: (pager: Pager) => void
   href?: (page: number) => string
 }